% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{dB.mode.per.row}
\alias{dB.mode.per.row}
\title{dB mode per row}
\usage{
dB.mode.per.row(x, windowSize = 5)
}
\arguments{
\item{x}{A numeric vector}

\item{windowSize}{An integer. Default = 5}
}
\value{
A numeric vector after some operation.
}
\description{
This function performs some operation.
}
