% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppRollMeandB-package.R
\docType{package}
\name{RcppRollMeandB-package}
\alias{RcppRollMeandB}
\alias{RcppRollMeandB-package}
\title{RcppRollMeandB: Wrapper package to calculate rolling statistics of acoustic data}
\description{
Wrapper package to calculate rolling statistics of acoustic data
}
\author{
\strong{Maintainer}: Johan van den Hoogen \email{jhoogen@ethz.ch}

}
\keyword{internal}
