% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{roll_meandB_threshold_wrap}
\alias{roll_meandB_threshold_wrap}
\title{Roll Min and Threshold}
\usage{
roll_meandB_threshold_wrap(
  x,
  windowRowSize = 9,
  windowColSize = 3,
  threshold = 3
)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{windowRowSize}{An integer.}

\item{windowColSize}{An integer.}

\item{threshold}{An integer.}
}
\value{
A numeric vector or matrix after some operation.
}
\description{
This function performs some operation.
}
