% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{roll_meandB_vector_wrap}
\alias{roll_meandB_vector_wrap}
\title{Rolling Mean dB Vector}
\usage{
roll_meandB_vector_wrap(x, windowSize = 3)
}
\arguments{
\item{x}{A numeric vector.}

\item{windowSize}{An integer.}
}
\value{
A numeric vector with the rolling mean dB
}
\description{
This function calculates the rolling mean dB for a vector.
}
