% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{roll_meandB_wrap}
\alias{roll_meandB_wrap}
\title{Rolling Mean dB}
\usage{
roll_meandB_wrap(x, windowSize = 3)
}
\arguments{
\item{x}{A numeric matrix.}

\item{windowSize}{An integer.}
}
\value{
A numeric matrix with the rolling mean dB
}
\description{
This function calculates the rolling mean dB
}
